Declare counter variables globally
Had to move the player and computer count variables outside of the playRound function
as they were getting reset each loop.  Couldn't figure out a way to include them inside
the game function but they function well in global scope.  Not sure if this is best
practice but the code is executing well