Clean code and add score keeping text to click event
I changed all the nested conditional statements to more concise && statements which cleared up
the code considerably.  I then was able to insert text content of the game score and add
a conditional to end the game.  Also was able to disable the buttons once the winning score is
reached.